Main.cpp =============================== Modified the main() function to have a game stage.
GameStage.h
===============================
Added a game() function declaration to the gameStage.h and gameStage.cpp.

Maps.h
===============================
Added a new header and implementation file "maps" for functions that work with the map files.
Added the function for parsing and loading the map files into memory.
Added the function for rendering the parsed map files.